+++ test/white/sample-test.js
@@ -1,0 +1,5 @@
+describe('Sample test', function() {
+  it('2 + 2 = 4', function() {
+    expect(2 + 2).to.equal(4);
+  });
+});