+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### WIP
+### 0.4.0 (November 4, 2013)
 - Convert custom Cake tasks into standard Jake tasks
 - Remove any prompts in favor of Jake arguments
 - Remove `npm run-script` tasks. Use `jake`.